requirement.txt
certifi==2022.12.7
Django==4.2.9
requests==2.31.0
python-decouple==3.8
